indexer/requirements.txt
web3==6.15.0
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.26.0
